Gear Up: Latest dev to stable.
[git-p4: depot-paths = "//depot/traktor/": change = 16235]

+++ code/Flash/FlashShape.cpp
@@ -54,23 +54,23 @@ bool FlashShape::create(const Aabb2& shapeBounds, const SwfShape* shape, const S
 		{
 			const SwfStraightEdgeRecord& s = shapeRecord->edge.straightEdge;
 			if (s.generalLineFlag)
-				path.lineTo(float(s.deltaX), float(s.deltaY), Path::CmRelative);
+				path.lineTo(s.deltaX, s.deltaY, Path::CmRelative);
 			else
 			{
 				if (!s.vertLineFlag)
-					path.lineTo(float(s.deltaX), 0.0f, Path::CmRelative);
+					path.lineTo(s.deltaX, 0, Path::CmRelative);
 				else
-					path.lineTo(0.0f, float(s.deltaY), Path::CmRelative);
+					path.lineTo(0, s.deltaY, Path::CmRelative);
 			}
 		}
 		else if (shapeRecord->edgeFlag && !shapeRecord->edge.straightFlag)
 		{
 			const SwfCurvedEdgeRecord& c = shapeRecord->edge.curvedEdge;
 			path.quadraticTo(
-				float(c.controlDeltaX),
-				float(c.controlDeltaY),
-				float(c.controlDeltaX + c.anchorDeltaX),
-				float(c.controlDeltaY + c.anchorDeltaY),
+				c.controlDeltaX,
+				c.controlDeltaY,
+				c.controlDeltaX + c.anchorDeltaX,
+				c.controlDeltaY + c.anchorDeltaY,
 				Path::CmRelative
 			);
 		}
@@ -87,8 +87,8 @@ bool FlashShape::create(const Aabb2& shapeBounds, const SwfShape* shape, const S
 			if (s.stateMoveTo)
 			{
 				path.moveTo(
-					float(s.moveDeltaX),
-					float(s.moveDeltaY),
+					s.moveDeltaX,
+					s.moveDeltaY,
 					Path::CmAbsolute
 				);
 			}
@@ -163,23 +163,23 @@ bool FlashShape::create(const SwfShape* shape)
 		{
 			const SwfStraightEdgeRecord& s = shapeRecord->edge.straightEdge;
 			if (s.generalLineFlag)
-				path.lineTo(float(s.deltaX), float(s.deltaY), Path::CmRelative);
+				path.lineTo(s.deltaX, s.deltaY, Path::CmRelative);
 			else
 			{
 				if (!s.vertLineFlag)
-					path.lineTo(float(s.deltaX), 0.0f, Path::CmRelative);
+					path.lineTo(s.deltaX, 0, Path::CmRelative);
 				else
-					path.lineTo(0.0f, float(s.deltaY), Path::CmRelative);
+					path.lineTo(0, s.deltaY, Path::CmRelative);
 			}
 		}
 		else if (shapeRecord->edgeFlag && !shapeRecord->edge.straightFlag)
 		{
 			const SwfCurvedEdgeRecord& c = shapeRecord->edge.curvedEdge;
 			path.quadraticTo(
-				float(c.controlDeltaX),
-				float(c.controlDeltaY),
-				float(c.controlDeltaX + c.anchorDeltaX),
-				float(c.controlDeltaY + c.anchorDeltaY),
+				c.controlDeltaX,
+				c.controlDeltaY,
+				c.controlDeltaX + c.anchorDeltaX,
+				c.controlDeltaY + c.anchorDeltaY,
 				Path::CmRelative
 			);
 		}
@@ -192,8 +192,8 @@ bool FlashShape::create(const SwfShape* shape)
 			if (s.stateMoveTo)
 			{
 				path.moveTo(
-					float(s.moveDeltaX),
-					float(s.moveDeltaY),
+					s.moveDeltaX,
+					s.moveDeltaY,
 					Path::CmAbsolute
 				);
 			}
@@ -225,7 +225,7 @@ bool FlashShape::create(const SwfShape* shape)
 	return true;
 }
 
-bool FlashShape::create(uint16_t fillBitmap, float width, float height)
+bool FlashShape::create(uint16_t fillBitmap, int32_t width, int32_t height)
 {
 	m_fillStyles.resize(1);
 	m_fillStyles[0].create(fillBitmap, Matrix33(
@@ -235,11 +235,11 @@ bool FlashShape::create(uint16_t fillBitmap, float width, float height)
 	));
 
 	Path path;
-	path.moveTo(0.0f, 0.0f, Path::CmAbsolute);
-	path.lineTo(width, 0.0f, Path::CmAbsolute);
+	path.moveTo(0, 0, Path::CmAbsolute);
+	path.lineTo(width, 0, Path::CmAbsolute);
 	path.lineTo(width, height, Path::CmAbsolute);
-	path.lineTo(0.0f, height, Path::CmAbsolute);
-	path.lineTo(0.0f, 0.0f, Path::CmAbsolute);
+	path.lineTo(0, height, Path::CmAbsolute);
+	path.lineTo(0, 0, Path::CmAbsolute);
 	path.end(1, 1, 0);
 
 	m_shapeBounds = path.getBounds();
